backend/requirements.txt
PyPDF2
langchain
google-generativeai
chromadb
python-dotenv
PyMuPDF
Pillow
Flask 
Flask-Cors